# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Uzbekistan OS API
  version: 0.1.0
  description: >-
    Contract skeleton for the Uzbekistan OS public API. Paths marked with
    x-implementation-status: planned are not implemented yet.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security: []
...
paths:
...
  /assistant/answer:
    post:
      tags: [assistant]
      operationId: answerGroundedQuestion
      summary: Answer from eligible official evidence with a bounded live fallback
      x-authorization:
        mode: authenticated
        roles: []
      x-idempotency:
        mode: none
        reason: The endpoint performs bounded generation and does not mutate application state.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantAnswerRequest"
      responses:
        "200":
          description: Grounded local or live-source answer, or safe insufficiency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantAnswerResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        Token verified by a configured trusted identity adapter. The API never
        trusts principal IDs or roles supplied directly by the caller.
...
  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    LanguageCode:
      type: string
      enum: [en, uz, ru]
...
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: [string, "null"]
        next_cursor:
          type: [string, "null"]
      additionalProperties: false
...
    RetrievalIntent:
      type: string
      enum:
        - arrival_entry
        - visa_eligibility
        - foreigner_registration
        - moving
        - business_registration
        - residence_permit
        - employment
        - study
        - banking
        - pinfl
        - healthcare
        - renting
        - customs
        - stay_extension
        - departure
        - general
    ConversationField:
      type: string
      enum:
        - audience
        - nationality
        - residency_status
        - location
        - accommodation_type
        - arrival_date
        - length_of_stay
        - applicant_role
...
    AnswerClaimCitation:
      type: object
      required: [evidence_id, quote]
      properties:
        evidence_id: { type: string, minLength: 1 }
        quote: { type: string, minLength: 3, maxLength: 2000 }
      additionalProperties: false
    AnswerClaim:
      type: object
      required: [id, text, citations]
      properties:
        id: { type: string, pattern: "^claim-[a-z0-9-]{1,63}$" }
        text: { type: string, minLength: 3, maxLength: 2000 }
        citations:
          type: array
          minItems: 1
          maxItems: 8
          items: { $ref: "#/components/schemas/AnswerClaimCitation" }
      additionalProperties: false
    GroundedAnswerSection:
      type: object
      required: [id, heading, claims]
      properties:
        id: { type: string, pattern: "^section-[a-z0-9-]{1,63}$" }
        heading: { type: string, minLength: 1, maxLength: 160 }
        claims:
          type: array
          minItems: 1
          maxItems: 24
          items: { $ref: "#/components/schemas/AnswerClaim" }
      additionalProperties: false
    ContextUsedItem:
      type: object
      required: [field, value, source_message_id]
      properties:
        field: { $ref: "#/components/schemas/ConversationField" }
        value: { type: string, minLength: 1, maxLength: 240 }
        source_message_id:
          { type: string, pattern: "^message-[a-z0-9-]{1,80}$" }
      additionalProperties: false
    ClarificationOption:
      type: object
      required: [value, label]
      properties:
        value: { type: string, minLength: 1, maxLength: 120 }
        label: { type: string, minLength: 1, maxLength: 160 }
      additionalProperties: false
    ClarificationRequest:
      type: object
      required: [id, field, question, reason, response_type, options]
      properties:
        id: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        field: { $ref: "#/components/schemas/ConversationField" }
        question: { type: string, minLength: 3, maxLength: 500 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        response_type: { type: string, enum: [single_choice, text] }
        options:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/ClarificationOption" }
      additionalProperties: false
    AnswerLimitation:
      type: object
      required: [code, message, evidence_ids]
      properties:
        code:
          type: string
          enum:
            - source_missing_detail
            - applicability_uncertain
            - source_conflict
            - context_missing
        message: { type: string, minLength: 3, maxLength: 500 }
        evidence_ids:
          type: array
          maxItems: 8
          uniqueItems: true
          items: { type: string }
      additionalProperties: false
    NextAction:
      type: object
      required: [id, type, label]
      properties:
        id: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        type: { type: string, enum: [continue_flow, change_context, view_evidence] }
        label: { type: string, minLength: 1, maxLength: 160 }
        workflow: { oneOf: [{ $ref: "#/components/schemas/RetrievalIntent" }, { type: "null" }] }
        field: { oneOf: [{ $ref: "#/components/schemas/ConversationField" }, { type: "null" }] }
      additionalProperties: false
    GroundedAnswer:
      type: object
      required:
        - status
        - language
        - summary
        - sections
        - context_used
        - limitations
        - next_actions
      properties:
        status: { type: string, enum: [answered, needs_clarification, insufficient] }
        language: { $ref: "#/components/schemas/LanguageCode" }
        summary: { type: string, minLength: 1, maxLength: 2000 }
        sections:
          type: array
          maxItems: 12
          items: { $ref: "#/components/schemas/GroundedAnswerSection" }
        clarification:
          oneOf:
            - { $ref: "#/components/schemas/ClarificationRequest" }
            - { type: "null" }
        context_used:
          type: array
          maxItems: 16
          items: { $ref: "#/components/schemas/ContextUsedItem" }
        limitations:
          type: array
          maxItems: 12
          items: { $ref: "#/components/schemas/AnswerLimitation" }
        next_actions:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/NextAction" }
      additionalProperties: false
    AssistantMessage:
      type: object
      required: [role, content]
      properties:
        role: { type: string, enum: [user, assistant] }
        content: { type: string, minLength: 1, maxLength: 2000 }
      additionalProperties: false
    ApplicabilityContext:
      type: object
      properties:
        audience: { type: [string, "null"], minLength: 1, maxLength: 120 }
        nationality: { type: [string, "null"], pattern: "^[A-Z]{2}$" }
        residency_status: { type: [string, "null"], minLength: 1, maxLength: 120 }
        location: { type: [string, "null"], minLength: 1, maxLength: 160 }
      additionalProperties: false
    AssistantAnswerRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 24
          items: { $ref: "#/components/schemas/AssistantMessage" }
        language: { type: [string, "null"], enum: [en, uz, ru, null] }
        applicability: { $ref: "#/components/schemas/ApplicabilityContext" }
      additionalProperties: false
    CitationReference:
      type: object
      required: [source_id, locator]
      properties:
        source_id: { type: string, format: uuid }
        locator: { type: string, minLength: 1 }
        quote: { type: [string, "null"] }
        source_url: { type: [string, "null"], format: uri }
        source_title: { type: [string, "null"] }
        reviewed_at: { type: [string, "null"], format: date-time }
      additionalProperties: false
    EvidenceItem:
      type: object
      required:
        - chunk_id
        - document_id
        - document_version_id
        - title
        - heading
        - content
        - content_hash
        - retrieval_score
        - citations
      properties:
        chunk_id: { type: string }
        document_id: { type: string }
        document_version_id: { type: string }
        title: { type: string }
        heading: { type: string }
        content: { type: string }
        content_hash: { type: string, pattern: "^[a-f0-9]{64}$" }
        retrieval_score: { type: number, minimum: 0, maximum: 1 }
        citations:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/CitationReference" }
      additionalProperties: false
    EvidencePack:
      type: object
      required:
        - plan_fingerprint
        - evidence_fingerprint
        - status
        - reason
        - total_characters
        - quarantined_chunk_ids
        - items
      properties:
        plan_fingerprint: { type: string }
        evidence_fingerprint: { type: string, pattern: "^[a-f0-9]{64}$" }
        status: { type: string, enum: [sufficient, insufficient] }
        reason: { type: [string, "null"] }
        total_characters: { type: integer, minimum: 0 }
        quarantined_chunk_ids:
          type: array
          items: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/EvidenceItem" }
      additionalProperties: false
    AssistantAnswerData:
      type: object
      required: [answer, evidence, intent, risk, accepted, issues, generated]
      properties:
        answer: { $ref: "#/components/schemas/GroundedAnswer" }
        evidence: { $ref: "#/components/schemas/EvidencePack" }
        intent: { type: string }
        risk: { type: string, enum: [medium, high] }
        accepted: { type: boolean }
        issues:
          type: array
          items: { type: string }
        generated: { type: boolean }
      additionalProperties: false
    AssistantAnswerResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AssistantAnswerData" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
      additionalProperties: false
...
    ErrorDetail:
      type: object
      required: [code, message, details]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
      additionalProperties: false
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
      additionalProperties: false